reorganizing runner library
